# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VinFast API Documentation",
    "version": "1.0.0",
    "description": "\n        ## 🚗 VinFast API Documentation\n\n        API documentation cho website cửa hàng ô tô VinFast - Hệ thống quản lý bán hàng và dịch vụ ô tô điện.\n\n        ### 🔐 Authentication\n        API sử dụng JWT (JSON Web Token) để xác thực. Để sử dụng các endpoint được bảo vệ:\n        1. Đăng nhập qua `/api/auth/login` để lấy token\n        2. Thêm token vào header: `Authorization: Bearer <token>`\n\n        ### 📱 Features\n        - **Authentication & Authorization**: Đăng ký, đăng nhập, quản lý người dùng\n        - **Car Management**: Quản lý xe, tìm kiếm, lọc theo loại\n        - **Order Management**: Đặt hàng, theo dõi đơn hàng, thanh toán\n        - **Test Drive**: Đăng ký lái thử, quản lý lịch hẹn\n        - **Contact**: Hệ thống liên hệ tư vấn\n        - **News**: Quản lý tin tức, bài viết\n        - **User Management**: Quản lý người dùng (Admin)\n\n        ### 🌐 Base URL\n        - Development: `http://localhost:8000/api`\n        - Production: `https://api.vinfast.com/api`\n      ",
    "contact": {
      "name": "VinFast Development Team",
      "email": "dev@vinfast.com",
      "url": "https://vinfast.vn"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token để xác thực. Format: Bearer <token>"
      }
    },
    "schemas": {
...
      "News": {
        "type": "object",
        "required": [
          "title",
          "excerpt",
          "content",
          "category",
          "featuredImage"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID bài viết"
          },
          "title": {
            "type": "string",
            "description": "Tiêu đề"
          },
          "slug": {
            "type": "string",
            "description": "Slug URL"
          },
          "excerpt": {
            "type": "string",
            "description": "Tóm tắt"
          },
          "content": {
            "type": "string",
            "description": "Nội dung"
          },
          "category": {
            "type": "string",
            "enum": [
              "product",
              "technology",
              "business",
              "event",
              "announcement",
              "review"
            ],
            "description": "Danh mục"
          },
          "featuredImage": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string"
              },
              "alt": {
                "type": "string"
              }
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "example": "published"
          },
          "featured": {
            "type": "boolean",
            "description": "Tin tức nổi bật"
          },
          "views": {
            "type": "number",
            "description": "Lượt xem"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Thông báo lỗi"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Chi tiết lỗi validation"
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Thông báo thành công",
            "example": "Thành công"
          },
          "data": {
            "type": "object",
            "description": "Dữ liệu trả về"
          },
          "count": {
            "type": "number",
            "description": "Số lượng items trong trang hiện tại",
            "example": 10
          },
          "total": {
            "type": "number",
            "description": "Tổng số items",
            "example": 100
          },
          "pagination": {
            "type": "object",
            "properties": {
              "next": {
                "type": "object",
                "properties": {
                  "page": {
                    "type": "number",
                    "example": 2
                  },
                  "limit": {
                    "type": "number",
                    "example": 10
                  }
                }
              },
              "prev": {
                "type": "object",
                "properties": {
                  "page": {
                    "type": "number",
                    "example": 1
                  },
                  "limit": {
                    "type": "number",
                    "example": 10
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/news": {
      "get": {
        "summary": "Lấy danh sách tin tức",
        "tags": [
          "News"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Số lượng tin tức mỗi trang"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "enum": [
                "product",
                "technology",
                "business",
                "event",
                "announcement",
                "review"
              ]
            },
            "description": "Lọc theo danh mục"
          },
          {
            "in": "query",
            "name": "featured",
            "schema": {
              "type": "boolean"
            },
            "description": "Lọc tin tức nổi bật"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách tin tức",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/News"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/news/featured": {
      "get": {
        "summary": "Lấy danh sách tin tức nổi bật",
        "tags": [
          "News"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Số lượng tin tức"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách tin tức nổi bật",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/News"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/news/search": {
      "get": {
        "summary": "Tìm kiếm tin tức",
        "tags": [
          "News"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            },
            "description": "Từ khóa tìm kiếm",
            "example": "VinFast"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "enum": [
                "product",
                "technology",
                "business",
                "event",
                "announcement",
                "review"
              ]
            },
            "description": "Lọc theo danh mục"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Số lượng tin tức mỗi trang"
          }
        ],
        "responses": {
          "200": {
            "description": "Kết quả tìm kiếm",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/News"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/news/{slug}": {
      "get": {
        "summary": "Lấy chi tiết tin tức",
        "tags": [
          "News"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Slug của tin tức",
            "example": "vinfast-vf-3-chinh-thuc-ra-mat"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết tin tức",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/News"
                            },
                            {
                              "type": "object",
                              "properties": {
                                "relatedPosts": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/News"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy tin tức",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}